requirements.txt
streamlit==1.23.1
openai==1.6.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2